# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Workforce OS v2 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings/org:
    get:
      operationId: getOrgSettings
      tags: [settings]
      summary: Get organization settings and compliance state
      responses:
        "200":
          description: Organization settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgSettings"
    put:
      operationId: updateOrgSettings
      tags: [settings]
      summary: Update organization settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrgInput"
      responses:
        "200":
          description: Updated org settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgSettings"
...
  /settings/icp:
    get:
      operationId: getIcpProfile
      tags: [settings]
      summary: Get ICP profile
      responses:
        "200":
          description: ICP profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IcpProfile"
    put:
      operationId: updateIcpProfile
      tags: [settings]
      summary: Update ICP profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IcpProfile"
      responses:
        "200":
          description: Updated ICP profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IcpProfile"
...
  /settings/integrations/{provider}/connect:
    post:
      operationId: connectIntegration
      tags: [settings]
      summary: Connect an integration (stub OAuth)
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Integration"
  /settings/integrations/{provider}/disconnect:
    post:
      operationId: disconnectIntegration
      tags: [settings]
      summary: Disconnect an integration
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Integration"
  /settings/cadence:
    get:
      operationId: getCadence
      tags: [settings]
      summary: Get sequence cadence config
      responses:
        "200":
          description: Cadence stages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CadenceStage"
    put:
      operationId: updateCadence
      tags: [settings]
      summary: Update sequence cadence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCadenceInput"
      responses:
        "200":
          description: Updated cadence
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CadenceStage"
  /settings/style:
    get:
      operationId: getStyleConfig
      tags: [settings]
      summary: Get brand voice and style config
      responses:
        "200":
          description: Style config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StyleConfig"
    put:
      operationId: updateStyleConfig
      tags: [settings]
      summary: Update brand voice config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StyleConfig"
      responses:
        "200":
          description: Updated style
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StyleConfig"
...
  /settings/team/invite:
    post:
      operationId: inviteTeamMember
      tags: [settings]
      summary: Invite a team member by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteInput"
      responses:
        "200":
          description: Invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /settings/team/{userId}:
    delete:
      operationId: removeTeamMember
      tags: [settings]
      summary: Remove a team member
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkActionResult"
...
  /settings/api-keys:
    get:
      operationId: listApiKeys
      tags: [settings]
      summary: List API keys (prefix only)
      responses:
        "200":
          description: API key list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
    post:
      operationId: createApiKey
      tags: [settings]
      summary: Issue a new API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyInput"
      responses:
        "201":
          description: Created key (full value shown once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyCreated"
  /settings/api-keys/{id}:
    delete:
      operationId: revokeApiKey
      tags: [settings]
      summary: Revoke an API key
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkActionResult"
  /settings/notifications:
    get:
      operationId: getNotificationPrefs
      tags: [settings]
      summary: Get notification preferences
      responses:
        "200":
          description: Notification prefs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPrefs"
    put:
      operationId: updateNotificationPrefs
      tags: [settings]
      summary: Update notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationPrefs"
      responses:
        "200":
          description: Updated prefs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPrefs"
...
components:
  schemas:
...
    BulkActionResult:
      type: object
      properties:
        affected: { type: integer }
      required: [affected]
...
    OrgSettings:
      type: object
      properties:
        orgId: { type: string }
        orgName: { type: string }
        slug: { type: string }
        logoUrl: { type: ["string", "null"] }
        country: { type: string }
        timezone: { type: string }
        senderName: { type: ["string", "null"] }
        liveSendEnabled: { type: boolean }
        postalAddress: { type: ["string", "null"] }
        unsubscribeUrl: { type: ["string", "null"] }
        suppressionCount: { type: integer }
        allowlistedDomains:
          type: array
          items: { type: string }
        plan: { type: string }
        creditsRemaining: { type: integer }
        welcomeComplete: { type: boolean }
      required: [orgId, orgName, slug, country, timezone, liveSendEnabled, suppressionCount, allowlistedDomains, creditsRemaining, welcomeComplete]
    UpdateOrgInput:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        logoUrl: { type: string }
        country: { type: string }
        timezone: { type: string }
        senderName: { type: string }
        postalAddress: { type: string }
        liveSendEnabled: { type: boolean }
...
    IcpProfile:
      type: object
      properties:
        industries:
          type: array
          items: { type: string }
        titles:
          type: array
          items: { type: string }
        geos:
          type: array
          items: { type: string }
        sizeBand: { type: string }
        intentSignals:
          type: array
          items: { type: string }
        seedDomains:
          type: array
          items: { type: string }
        exclusionDomains:
          type: array
          items: { type: string }
      required: [industries, titles, geos, sizeBand, intentSignals, seedDomains, exclusionDomains]
    Integration:
      type: object
      properties:
        id: { type: string }
        provider: { type: string }
        status:
          type: string
          enum: [connected, available, errored]
        accountEmail: { type: ["string", "null"] }
        connectedAt: { type: ["string", "null"] }
        errorMessage: { type: ["string", "null"] }
      required: [id, provider, status]
    CadenceStage:
      type: object
      properties:
        id: { type: string }
        dayOffset: { type: integer }
        channel: { type: string }
        label: { type: string }
        enabled: { type: boolean }
        position: { type: integer }
      required: [id, dayOffset, channel, label, enabled, position]
    StyleConfig:
      type: object
      properties:
        voice: { type: string }
        toneValue: { type: integer }
        signatureHtml: { type: string }
      required: [voice, toneValue, signatureHtml]
    TeamMember:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        role:
          type: string
          enum: [OWNER, ADMIN, MEMBER]
        status: { type: string }
        invitedAt: { type: string }
        joinedAt: { type: ["string", "null"] }
      required: [id, email, name, role, status, invitedAt]
    InviteInput:
      type: object
      properties:
        email: { type: string }
        name: { type: string }
        role:
          type: string
          enum: [ADMIN, MEMBER]
      required: [email, name, role]
...
    ApiKey:
      type: object
      properties:
        id: { type: string }
        prefix: { type: string }
        name: { type: string }
        lastUsedAt: { type: ["string", "null"] }
        createdAt: { type: string }
      required: [id, prefix, name, createdAt]
    CreateApiKeyInput:
      type: object
      properties:
        name: { type: string }
      required: [name]
    ApiKeyCreated:
      type: object
      properties:
        id: { type: string }
        prefix: { type: string }
        name: { type: string }
        fullKey: { type: string }
        createdAt: { type: string }
      required: [id, prefix, name, fullKey, createdAt]
    NotificationPrefs:
      type: object
      properties:
        emailEnabled: { type: boolean }
        slackEnabled: { type: boolean }
        approvalQueueFull: { type: boolean }
        sendFailed: { type: boolean }
        suppressionHit: { type: boolean }
        weeklyReport: { type: boolean }
        newReply: { type: boolean }
      required: [emailEnabled, slackEnabled, approvalQueueFull, sendFailed, suppressionHit, weeklyReport, newReply]
...
    UpdateCadenceInput:
      type: object
      properties:
        stages:
          type: array
          items:
            $ref: "#/components/schemas/CadenceStage"
      required: [stages]
...