# lib/api-zod/src/generated/types/cadenceStage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Workforce OS v2 API
 * OpenAPI spec version: 0.1.0
 */

export interface CadenceStage {
  id: string;
  dayOffset: number;
  channel: string;
  label: string;
  enabled: boolean;
  position: number;
}