# lib/api-zod/src/generated/types/notificationPrefs.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Workforce OS v2 API
 * OpenAPI spec version: 0.1.0
 */

export interface NotificationPrefs {
  emailEnabled: boolean;
  slackEnabled: boolean;
  approvalQueueFull: boolean;
  sendFailed: boolean;
  suppressionHit: boolean;
  weeklyReport: boolean;
  newReply: boolean;
}