# lib/api-zod/src/generated/types/updateOrgInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Workforce OS v2 API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateOrgInput {
  name?: string;
  slug?: string;
  logoUrl?: string;
  country?: string;
  timezone?: string;
  senderName?: string;
  postalAddress?: string;
  liveSendEnabled?: boolean;
}